requirements.txt
Django==3.2.14
djangorestframework==3.13.1
django-filter==22.1
markdown==3.4.1
djangorestframework_simplejwt==5.2.0